add Bloco + fix id type

+++ src/main/java/com/devsuperior/dscommerce/entities/Categoria.java
@@ -8,7 +8,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(columnDefinition = "TEXT")
     private String descricao;
@@ -16,16 +16,16 @@ public class Categoria {
     public Categoria() {
     }
 
-    public Categoria(Long id, String descricao) {
+    public Categoria(Integer id, String descricao) {
         this.id = id;
         this.descricao = descricao;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
